hybrid-nmt/requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
sentencepiece>=0.1.99
sacrebleu>=2.3.1
accelerate>=0.20.0
indic-nlp-library>=0.95
unbabel-comet>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
huggingface-hub>=0.16.0
